Enrichment analysis of go

run enrich_go and plot results

enrichGO_run(
  deg_data,
  x,
  y,
  cut_FC = 1,
  cut_P = 0.05,
  top = 10,
  dir = ".",
  prefix = "3-EnrichGO",
  OrgDb = "org.Hs.eg.db",
  keyType = "SYMBOL",
  ont = "ALL",
  simplify = TRUE,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  label = c("Down", "Stable", "Up"),
  label_ns = "Stable",
  mc.cores = 1L
)

Arguments

deg_data
a DEG data frame contains logFC and p value

x
which column is log FC

y
which column is P value

cut_FC
a single number character or numeric vector in threshold value of log FC

cut_P
threshold value of P value, can set for every cut_FC number in numeric vector format

top
top rows

dir
where to save results files

prefix
a prefix of file names in this step

OrgDb
OrgDb

keyType
keytype of input gene

ont
One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.

simplify
whether simplify

pvalueCutoff
adjusted pvalue cutoff on enrichment tests to report

pAdjustMethod
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

qvalueCutoff
qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.

minGSSize
minimal size of genes annotated by Ontology term for testing.

maxGSSize
maximal size of genes annotated for testing

readable
whether mapping gene ID to gene Name

pool
If ont='ALL', whether pool 3 GO sub-ontologies

label
symbol word for groups

label_ns
which group is the stable group

mc.cores
param for mclapply, choose cores to use in non-Windows machine

Value

a list result files of GO

Examples

if (FALSE) {
enrichGO_run(DEG_df, x = "log2FoldChange", y = "pvalue", dir = tempdir())
}