enrich_go

a funtion group deg data frame and make enrichment analysis in one step

enrich_kegg(
  deg_data,
  x,
  y,
  cut_FC = 1,
  cut_P = 0.05,
  organism = "hsa",
  keyType = "kegg",
  OrgDb = "org.Hs.eg.db",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  use_internal_data = FALSE,
  label = c("Down", "Stable", "Up"),
  label_ns = "Stable",
  mc.cores = 1L
)

Arguments

deg_data

a DEG data frame contains logFC and p value

x

which column is log FC

y

which column is P value

cut_FC

a single number character or numeric vector in threshold value of log FC

cut_P

threshold value of P value, can set for every cut_FC number in numeric vector format

organism

supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'

keyType

one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'

OrgDb

OrgDb

pvalueCutoff

adjusted pvalue cutoff on enrichment tests to report

pAdjustMethod

one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

minGSSize

minimal size of genes annotated by Ontology term for testing.

maxGSSize

maximal size of genes annotated for testing

qvalueCutoff

qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.

use_internal_data

logical, use KEGG.db or latest online KEGG data

label

symbol word for groups

label_ns

which group is the stable group

mc.cores

param for mclapply, choose cores to use in non-Windows machine

Value

enrichKEGG result

Examples

if (FALSE) {
enrich_kegg(deg_data = DEG_df, x = "log2FoldChange", y = "pvalue")
}