Pre-check you data by PCA and Heat maps

A previous check function for overview the data sets.

pre_check(
  counts_data,
  group_list,
  dir = ".",
  prefix = "1-pre_check",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)

Arguments

counts_data
a counts data frame of rows in genes and columns in samples

group_list
a list ordered by samples in counts_data

dir
a directory to store results

prefix
a prefix of file names in this step

palette
a color palette for plots

Value

a directory contains figures for data QC check

Examples

pre_check(counts_input, group_list, tempdir())
#> Too few points to calculate an ellipse
#> Too few points to calculate an ellipse
#> 🌳 PCA checking have done, a plot was store in /tmp/RtmpegxAnx.
#> 🌳 Correlation checking have done, a plot was store in /tmp/RtmpegxAnx.
#> 🌳 Correlation to top 500 genes checking have done, a plot was store in /tmp/RtmpegxAnx.
#> 🌳 Standard Deviation top 1000 genes checking have done, a plot was store in /tmp/RtmpegxAnx.