Choose top genes from DEG data frame

For example, if top = 50, head 50 and tail 50 will return from a ordered DEG data frame after filtering by threshold value of P value

top_deg(deg_data, x, y, top, cut_P = 0.05)

Arguments

deg_data
a DEG data frame contains logFC and p value

x
which column is log FC

y
which column is P value

top
a single number or a length of 2 numeric vector, if 2 numeric vector, first one is top max logFC.

cut_P
a single number for threshold value of P value

Value

a character vector of top genes

Examples

top_deg(DEG_df, x = "log2FoldChange", y = "pvalue", 50, 0.05)

#>   [1] "AP001066.1"  "AC021534.1"  "AC140658.1"  "VWA7"        "AL356489.4" 
#>   [6] "LINC02593"   "TRIM63"      "NRN1"        "LINC01844"   "HSD3B2"     
#>  [11] "SMIM10L2A"   "SNAI3"       "FER1L4"      "SLC14A2"     "SLC7A10"    
#>  [16] "AL356414.1"  "SLC7A8"      "SEC14L5"     "PLA2G4D"     "WNT6"       
#>  [21] "AL355864.1"  "NEURL3"      "SAMD11"      "STC1"        "AC108451.2" 
#>  [26] "ASB4"        "AC079466.1"  "CBFA2T3"     "BDKRB2"      "TMEM51-AS1" 
#>  [31] "CACNA1H"     "NAV2-AS4"    "CACNA1B"     "PRUNE2"      "NECTIN4-AS1"
#>  [36] "LINC02412"   "LINC01050"   "UPK1A"       "NOVA2"       "SLC6A12"    
#>  [41] "PWWP3B"      "MIR210HG"    "AC018470.1"  "AC117500.3"  "PPP1R16B"   
#>  [46] "SERPINA6"    "AC117500.5"  "AP000695.1"  "CA9"         "TREM1"      
#>  [51] "ANKRD1"      "LINC01488"   "RTL3"        "MARCHF4"     "AL356289.2" 
#>  [56] "AC100800.1"  "ANKRD2"      "KMO"         "AC245041.1"  "ANXA3"      
#>  [61] "AC245041.2"  "AC087269.3"  "LINC01776"   "C9orf135"    "NPPB"       
#>  [66] "KRTAP4-8"    "EDN1"        "LYPD6B"      "TSPAN2"      "LINC02383"  
#>  [71] "OLR1"        "AL592211.2"  "LINC01186"   "AC110792.4"  "AP003555.1" 
#>  [76] "LINC02533"   "KRTAP2-3"    "PLCE1-AS1"   "CPA4"        "CCN2"       
#>  [81] "AF015262.1"  "AC093732.1"  "CDH1"        "FLJ30679"    "AC025154.2" 
#>  [86] "RBM24"       "IL6"         "LINC00052"   "AC062015.1"  "AL357833.1" 
#>  [91] "GSTM5"       "KRT8P14"     "MYPN"        "LINC01638"   "AC080011.1" 
#>  [96] "SHISA9"      "AP001525.1"  "PSG9"        "ITGBL1"      "NAV3"